% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotype_data}
\alias{phenotype_data}
\title{An example of phenotype data.}
\format{A numeric matrix or data.frame with three columns for three phenotypes
and 500 rows for the individuals in the sample.}
\usage{
data(phenotype_data)
}
\description{
Here phenotype\_data is a data.frame with three columns for three phenotypes
and the number of rows to be the number of individuals in the sample (500 in this toy data).
 Data for each phenotype provided must be adjusted individually for relevant
  covariates (e.g., age, sex, genetic ancestry) beforehand,
   and should follow a normal distribution.
}
\examples{
data(phenotype_data)
pheno <- phenotype_data
}
\keyword{datasets}
